api/requirements.txt
fastapi==0.73.0
motor==2.5.1
pydantic[email]
python-decouple==3.6
uvicorn==0.17.4
pandas==1.3.5
urllib3==1.21.1

# Chatbot
google-cloud-aiplatform==1.28.1
inflect
scikit-learn